Render all sectors without OpenGL.
This looks horrifyingly ugly, because you end
up with geometry poking through in the places
where the test map is non-euclidean.  And the
proportions of the projection are off.  But
the basics of the projection are in place,
and after some tweaks, it will be a good
jumping off point to doing selective sector
rendering.