Fixed FOV visibility testing.
The original version was a reminder that for
even simple 2d geometry, you can not gloss over
any detail. You forget any corner cases and
it will be painfully obvious.

o Fixed up FOV visibility debug drawing so it
  scales, and draws a background so I don't
  have to worry about contrast with the
  background.

o Fixed algorithm so it wouldn't mark things as behind
  you as visible.

o Changed from specifying FOVy to FOVx, as that seems
  more intuitive.

Note this doesn't make the drawing perfect.  For
sectors that overlap in 3d space, you can still
get artifacts where an overlapping sector is drawn
when it should not be.  This is a deficiency in the
algorithm itself, since solid walls don't block visibility
when projecting the FOV through sector edges.

There's an obvious way to brute force that, but it
gets ugly without a spatial map to look up where
the solid walls go.  So that needs more thought.